use login logic to context
